moves primitives to package primitive
